+++ src/main/java/com/fluffygram/newsfeed/domain/user/repository/UserRepository.java
@@ -4,10 +4,13 @@ import com.fluffygram.newsfeed.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     default User findByIdOrElseThrow(Long id) {
         return findById(id).orElseThrow(()-> new RuntimeException("[id = " + id + "] 에 해당하는 유저가 존재하지 않습니다."));
     }
 
+    Optional<User> findByEmail(String email);
 }
