+++ src/main/java/com/fluffygram/newsfeed/domain/user/service/UserService.java
@@ -3,19 +3,18 @@ package com.fluffygram.newsfeed.domain.user.service;
 import com.fluffygram.newsfeed.domain.base.Valid.AccessWrongValid;
 import com.fluffygram.newsfeed.domain.user.dto.UserResponseDto;
 import com.fluffygram.newsfeed.domain.user.entity.User;
+import com.fluffygram.newsfeed.domain.user.image.GetUserImage;
+import com.fluffygram.newsfeed.domain.user.image.UploadUserImage;
 import com.fluffygram.newsfeed.domain.user.repository.UserRepository;
-import com.fluffygram.newsfeed.global.config.Const;
 import com.fluffygram.newsfeed.global.config.PasswordEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
-import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.net.MalformedURLException;
-import java.nio.file.Path;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +26,9 @@ public class UserService {
     private final AccessWrongValid accessWrongValid;
 
     public UserResponseDto signUp(String email,String password, String userNickname, String phoneNumber, MultipartFile profileImage) {
+        Optional<User> userByEmail = userRepository.findByEmail(email);
+        userByEmail.ifPresent(user -> {throw new RuntimeException("이미 존재하는 유저입니다.");});
+
         String encodedPassword = passwordEncoder.encode(password);
 
         String profileImageUrl = uploadUserImage.uploadUserImage(profileImage);
@@ -50,13 +52,15 @@ public class UserService {
     }
 
     @Transactional
-    public UserResponseDto updateUserById(Long id, String password, String userNickname, String phoneNumber, String profileImage) {
+    public UserResponseDto updateUserById(Long id, String password, String userNickname, String phoneNumber, MultipartFile profileImage) {
+        String profileImageUrl = uploadUserImage.uploadUserImage(profileImage);
+
         User user = userRepository.findByIdOrElseThrow(id);
 
         user.updatePassword(passwordEncoder.encode(password));
         user.updateUserNickname(userNickname);
         user.updatePhoneNumber(phoneNumber);
-        user.updateProfileImage(profileImage);
+        user.updateProfileImage(profileImageUrl);
 
         User savedUser = userRepository.save(user);
 
@@ -69,31 +73,16 @@ public class UserService {
 
         User user = userRepository.findByIdOrElseThrow(id);
 
-        if(passwordEncoder.matches(password, user.getPassword())){
-            throw new RuntimeException("비밀번호가 일치하지 않습니다.");
-        }
+//        if(passwordEncoder.matches(password, user.getPassword())){
+//            throw new RuntimeException("비밀번호가 일치하지 않습니다.");
+//        }
 
         userRepository.delete(user);
     }
 
     public Resource getUserImage(Long id) {
-        Resource resource;
         String imageUrl = userRepository.findByIdOrElseThrow(id).getProfileImage();
 
-        try {
-            // 파일 경로 생성
-            Path filePath = Const.IMAGE_STORAGE_PATH.resolve(imageUrl).normalize();
-
-            // 파일을 Resource 객체로 변환
-            resource = new UrlResource(filePath.toUri());
-            if (!resource.exists()) {
-                throw new RuntimeException("파일이 없습니다.");
-            }
-        }
-        catch (MalformedURLException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-
-        return resource;
+        return GetUserImage.getImage(imageUrl);
     }
 }
